requirements.txt
routeros-api==0.21.0
scapy==2.7.0
getmac==0.9.5
python-nmap==0.7.1
pysnmp==7.1.22
zeroconf==0.148.0
requests==2.32.5
fastapi==0.135.1
uvicorn==0.42.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==4.1.0
httpx==0.28.1
pydantic==2.12.5
starlette==0.52.1
fpdf2==2.8.7
jinja2==3.1.6
markupsafe==3.0.3
cryptography==46.0.5
defusedxml==0.7.1
fonttools==4.62.1
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
ply==3.11
pyasn1==0.4.8
pycparser==3.0
pycryptodomex==3.23.0
pysmi==1.6.3
typing_extensions==4.15.0
typing_inspection==0.4.2
ifaddr==0.2.0
anyio==4.12.1
idna==3.11
certifi==2026.2.25
charset_normalizer==3.4.6
h11==0.16.0
httpcore==1.0.9
iniconfig==2.3.0
urllib3==2.6.3
fastapi
uvicorn
psutil
google-genai
